lara/state/land: port M_Special to TRX
