+++ src/tr1/game/lara/common.h
@@ -11,9 +11,6 @@
 
 void Lara_Control(void);
 
-ITEM *Lara_GetDeathCameraTarget(void);
-void Lara_SetDeathCameraTarget(int16_t item_num);
-
 void Lara_TakeHit(ITEM *lara_item, int32_t x, int32_t z);
 void Lara_ControlExtra(int16_t item_num);
