--- src/tr1/game/lara/state.c
@@ -1,37 +1,0 @@
-#include "game/input.h"
-#include "game/lara/common.h"
-#include "game/lara/look.h"
-#include "game/objects/common.h"
-#include "game/objects/effects/twinkle.h"
-#include "game/sound.h"
-#include "global/vars.h"
-
-#include <libtrx/config.h>
-#include <libtrx/game/camera.h>
-#include <libtrx/game/lara.h>
-#include <libtrx/game/lara/util.h>
-#include <libtrx/game/math.h>
-
-#include <stdint.h>
-
-static void M_Special(ITEM *item, COLL_INFO *coll);
-
-static void M_Special(ITEM *item, COLL_INFO *coll)
-{
-    ITEM *const target_item = Lara_GetDeathCameraTarget();
-    if (target_item != nullptr) {
-        g_Camera.item = target_item;
-        g_Camera.flags = CF_CHASE_OBJECT;
-        g_Camera.type = CAM_FIXED;
-        g_Camera.target_angle = item->rot.y;
-        g_Camera.target_distance = CAM_SPECIAL_DISTANCE;
-    } else {
-        g_Camera.flags = CF_FOLLOW_CENTRE;
-        g_Camera.target_angle = CAM_SPECIAL_ANGLE;
-    }
-    g_Camera.target_elevation = CAM_SPECIAL_ELEVATION;
-}
-
-// clang-format off
-REGISTER_LARA_STATE(LS_SPECIAL,       M_Special)
-// clang-format on
