--- src/tr2/game/lara/state.c
@@ -1,29 +1,0 @@
-#include "game/input.h"
-#include "game/inventory.h"
-#include "game/lara/control.h"
-#include "game/lara/misc.h"
-#include "game/overlay.h"
-#include "game/sound.h"
-#include "game/viewport.h"
-#include "global/vars.h"
-
-#include <libtrx/config.h>
-#include <libtrx/game/camera.h>
-#include <libtrx/game/game.h>
-#include <libtrx/game/lara.h>
-#include <libtrx/game/lara/util.h>
-#include <libtrx/game/music.h>
-#include <libtrx/utils.h>
-
-static void M_Special(ITEM *item, COLL_INFO *coll);
-
-static void M_Special(ITEM *item, COLL_INFO *coll)
-{
-    g_Camera.flags = CF_FOLLOW_CENTRE;
-    g_Camera.target_angle = CAM_SPECIAL_ANGLE;
-    g_Camera.target_elevation = CAM_SPECIAL_ELEVATION;
-}
-
-// clang-format off
-REGISTER_LARA_STATE(LS_SPECIAL,      M_Special)
-// clang-format on
